requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow-cpu
tensorflow_hub
pandas
